+++ .env.example
@@ -1,0 +1,2 @@
+WEB_PORT=3000
+DOC_PORT=3001
