+++ apps/web/biome.json
@@ -1,6 +1,6 @@
 {
 	"extends": ["../../packages/tooling/biome/config.json"],
 	"files": {
-		"ignore": [".next/**"]
+		"ignore": [".next/**, .turbo/**"]
 	}
 }
