+++ package.json
@@ -14,6 +14,5 @@
 	"engines": {
 		"node": ">=18"
 	},
-	"packageManager": "npm@10.2.4",
 	"workspaces": ["apps/*", "packages/*"]
 }
