+++ packages/ui/package.json
@@ -14,6 +14,8 @@
 	"devDependencies": {
 		"@biomejs/biome": "^1.8.0",
 		"@if/tooling": "*",
+		"@storybook/react": "^8.1.6",
+		"@storybook/test": "^8.1.6",
 		"@turbo/gen": "^1.12.4",
 		"@types/eslint": "^8.56.5",
 		"@types/node": "^20.11.24",
