+++ packages/ui/src/Button/index.tsx
@@ -5,17 +5,18 @@ import type { ReactNode } from 'react';
 interface ButtonProps {
 	children: ReactNode;
 	className?: string;
-	appName: string;
 }
 
-export const Button = ({ children, className, appName }: ButtonProps) => {
+function Button({ children, className }: ButtonProps) {
 	return (
 		<button
 			type='button'
 			className={className}
-			onClick={() => alert(`Hello from your ${appName} app!`)}
+			onClick={() => alert('Hello world')}
 		>
 			{children}
 		</button>
 	);
-};
+}
+
+export default Button;
