+++ tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "@repo/tooling/tsconfig/base.json"
+	"extends": "@if/tooling/tsconfig/base.json"
 }
